Add files via upload
We have also improved the code as before whenever we typed a prompt, it was only able to download the file with 3 words meaning whenever someone typed in 4 words, it would download the image as 3 words and will be unable to load since the full text does not match the image file. We fixed it by changing the search prompt to instead the first 3 words with no text limit. (line 68 on feature_2.py from previous build)